requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
umap-learn>=0.5.4
hdbscan>=0.8.33
hypothesis>=6.82.0
pytest>=7.4.0
jupyter>=1.0.0
tqdm>=4.65.0
Pillow>=10.0.0
